+++ lib/l10n/app_localizations_cs.dart
@@ -4834,6 +4834,9 @@ class AppLocalizationsCs extends AppLocalizations {
   String get editMenuTitle => 'Upravit';
 
   @override
+  String get editorDiscardChanges => 'Zahodit změny';
+
+  @override
   String get editorInsertDivider => 'Vložit oddělovač';
 
   @override